Agentic_Systems/crewai/mahpad_flow/mahpad/pyproject.toml
[project]
name = "mahpad"
version = "0.1.0"
description = "mahpad using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.121.0,<1.0.0",
    "psycopg2-binary>=2.9.10",
]

[project.scripts]
kickoff = "mahpad.main:kickoff"
run_crew = "mahpad.main:kickoff"
plot = "mahpad.main:plot"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "flow"
